+++ src/commands/build-react.js
@@ -1,9 +1,11 @@
 import path from 'path'
 
+import resolve from 'resolve'
 import runSeries from 'run-series'
 
 import {UserError} from '../errors'
 import webpackBuild from '../webpackBuild'
+import {install} from '../utils'
 import cleanApp from './clean-app'
 
 // Using a config function as webpackBuild() sets NODE_ENV to production if it
@@ -12,51 +14,63 @@ import cleanApp from './clean-app'
 function buildConfig(args) {
   let entry = args._[1]
   let dist = args._[2] || 'dist'
+  let mountId = args['mount-id'] || 'app'
 
+  let basedir = process.cwd()
   let production = process.env.NODE_ENV === 'production'
   let filenamePattern = production ? '[name].[chunkhash:8].js' : '[name].js'
 
   let config = {
     babel: {
       commonJSInterop: true,
-      stage: 0,
       presets: ['react'],
+      stage: 0,
     },
     devtool: 'source-map',
     entry: {
-      app: [path.resolve(entry)]
+      // Use a dummy entry module to support rendering an exported React
+      // Component or Element for quick prototyping.
+      app: [require.resolve('../reactRunEntry')],
     },
     output: {
-      filename: filenamePattern,
       chunkFilename: filenamePattern,
+      filename: filenamePattern,
       path: path.resolve(dist),
       publicPath: '/',
     },
     plugins: {
+      define: {
+        NWB_REACT_RUN_MOUNT_ID: JSON.stringify(mountId)
+      },
       html: {
-        mountId: args['mount-id'] || 'app',
+        mountId,
         title: args.title || 'React App',
       },
       // A vendor bundle must be explicitly enabled with a --vendor flag
       vendor: args.vendor,
     },
+    resolve: {
+      alias: {
+        // Allow the dummy entry module to import the provided entry module
+        'nwb-react-run-entry': path.resolve(entry),
+        // Allow the dummy entry module to resolve React and ReactDOM from the cwd
+        'react': path.dirname(resolve.sync('react/package.json', {basedir})),
+        'react-dom': path.dirname(resolve.sync('react-dom/package.json', {basedir})),
+      }
+    }
+  }
+
+  if (args.polyfill === false || args.polyfills === false) {
+    config.polyfill = false
   }
 
   if (args.inferno) {
-    config.resolve = {
-      alias: {
-        'react': 'inferno-compat',
-        'react-dom': 'inferno-compat',
-      }
-    }
+    config.resolve.alias.react = config.resolve.alias['react-dom'] =
+      path.dirname(resolve.sync('inferno-compat/package.json', {basedir}))
   }
   else if (args.preact) {
-    config.resolve = {
-      alias: {
-        'react': 'preact-compat',
-        'react-dom': 'preact-compat',
-      }
-    }
+    config.resolve.alias.react = config.resolve.alias['react-dom'] =
+      path.dirname(resolve.sync('preact-compat/package.json', {basedir}))
   }
 
   if (production) {
@@ -77,10 +91,18 @@ export default function buildReact(args, cb) {
   let dist = args._[2] || 'dist'
 
   let library = 'React'
-  if (args.inferno) library = 'Inferno (React compat)'
-  else if (args.preact) library = 'Preact (React compat)'
+  let packages = ['react', 'react-dom']
+  if (args.inferno) {
+    library = 'Inferno (React compat)'
+    packages.push('inferno', 'inferno-compat')
+  }
+  if (args.preact) {
+    library = 'Preact (React compat)'
+    packages.push('preact', 'preact-compat')
+  }
 
   runSeries([
+    (cb) => install(packages, {check: true}, cb),
     (cb) => cleanApp({_: ['clean-app', dist]}, cb),
     (cb) => webpackBuild(`${library} app`, args, buildConfig, cb),
   ], cb)